Alter build matrix
This commit changes the build matrix in the TravisCI build config. Currently, RBX-2 doesn't support the **kwargs argument in methods. This functionality is merged into head, so we'll be testing against that for the time being. Also, ruby-2.0 refinements do not work the same as ruby-2.1, so we're disabling tests on that platform.